Implement per-canvas layouts:
A canvas has a vector of allowed layouts (instead of 'all' as
before). One of the layouts can be 0 to deactivate canvas. This allows
the specification of canvas-specific layouts, while retaining the
capability to use the same layout on multiple canvases.

Change all config files accordingly. Programming Guide updates to be done.

[ ] May break build
[ ] Breaks existing applications
[ ] Bugfix
[x] New Feature
[x] Cleanup
[ ] Optimization
[ ] Documentation