Fix (Martin Lambers):
The handling of the Shift key in libs/client/glXEventHandler.cpp
currently only gives upper case characters A-Z, but does not work for
other keys produced with the Shift key, such as ()<>!@#$% etc.

(Note that this fix still is not entirely complete: for that, it would
have to take the Lock modifier into account and various settings that
influence its use, see
<http://tronche.com/gui/x/xlib/input/keyboard-encoding.html>. But I
guess nobody ever uses this so it's not worth the trouble).

[ ] May break build
[ ] May break existing applications (see CHANGES.txt)
[x] Bugfix
[ ] New Feature
[ ] Cleanup
[ ] Optimization
[ ] Documentation