Remove use of RefPtr to connection in event thread.
If the application exits without explicitly closing the connection, the
refptr in the event thread keeps the instance from being deleted.  Since we
(believe we) guarantee that the event thread will exit on close() (since we
join it), and we call close() in the RDMAConnection destructor, we should never
have a problem with an invalid pointer.

[ ] May break build
[ ] May break existing applications (see CHANGES.txt)
[x] Bugfix
[ ] New Feature
[ ] Cleanup
[ ] Optimization
[ ] Documentation