Change compound default task computation, preparing the ROI implementation: * non-leafs now have a default clear task * leafs have only a clear task if no parent with the same channel has clear
Reasoning is that in any case the destination should clear, not the
source/dest compound since this is a sub-task of the dest. So far it did
not show up, since without ROI in 2D the dest is fully covered by the
input images, needing no clear. It did show up for dest-only DB
compounds, but nobody noticed.

Seems to be a step in the right direction, since it simplifies quite a
few config files.

[ ] May break build
[ ] May break existing applications (see CHANGES.txt)
[ ] Bugfix
[x] New Feature
[x] Cleanup
[ ] Optimization
[ ] Documentation