Implement per-window space mouse events (#150)
The spnav API registers and sends the event to windows, therefore the
per-node hack needed on AGL is unnecessary. Implement structure to track
which windows and which display has been used. This should take care of
multiple, dynamic windows per display, but fails if multiple dynamic
displays are used, since closing the origin Display will close the
connection to the spacenavd. Opened
https://sourceforge.net/tracker/index.php?func=detail&aid=3553508&group_id=200169&atid=972425#
to get API changed.