Removing deprecated usage of Config::sendEvent and Config::handleEvent.
This is also fixing an issue on client applications using only eq::Config::handleEvent(eq::EventICommand) to handle events. Somehow the event type obtained in that method was the packet size of the deprecated ConfigEvent packet instead of the actual event. This wasn't affecting eqPly and eVolve as they were both still using the pre 2.0 API.